long long?

+++ npbenchmark-main/SDK.PCP/pcp_vector.h
@@ -58,20 +58,20 @@ private: // member variables.
     int prev_num_uncovered;
 
     Move moved;
-    int min_delta;
+    long long int min_delta;
     int tabu_open;
     int tabu_close;
     // Dimension: num_vertex;
     // Meaning: element j is sum of the weights of the vertices which can only be served by center j.
     // Therefore, it can evaluate the importance of this center.
-    std::vector<int> center_weights;
+    std::vector<long long int> center_weights;
     // dimension: num_vertex;
     // meaning: element j is sum of the weights of the vertices which can only be served by center j.
-    std::vector<int> prev_center_weights;
+    std::vector<long long int> prev_center_weights;
     std::vector<Move> equal_pair;
     int equal_pair_count;
 
-    int sum_uncovered_weight; // f(X), 记录未覆盖顶点的加权和, Eq (11);
+    long long int sum_uncovered_weight; // f(X), 记录未覆盖顶点的加权和, Eq (11);
 
     long long iter;
